# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahara Triage Voice Agent API",
    "summary": "Conversational health-triage voice agent: Supabase-backed triage, scheduling, and clinician workflows.",
    "description": "Patient signup and voice triage (browser WebSocket or Africa's Talking phone channel), automatic doctor-slot routing, clinician queues with embedded triage reports, prescriptions, and consent-gated call workflows.\n\nAuth: Supabase session JWT as `Authorization: Bearer <token>` (use the Authorize button). The browser voice channel is a WebSocket and is documented in the README message catalog; phone callbacks accept `application/x-www-form-urlencoded` and return TwiML XML.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.sahara.example.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/prescriptions": {
      "get": {
        "tags": [
          "prescriptions"
        ],
        "summary": "My prescription inbox",
        "description": "Patient-facing prescription inbox.",
        "operationId": "list_my_prescriptions_api_prescriptions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "prescriptions"
        ],
        "summary": "Send a prescription to a patient",
        "description": "Clinician-facing: send a prescription to a patient.",
        "operationId": "create_prescription_api_prescriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionCreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Not a recognized patient session"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Prescription": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "medication": {
            "type": "string",
            "title": "Medication"
          },
          "dosage": {
            "type": "string",
            "title": "Dosage"
          },
          "instructions": {
            "type": "string",
            "title": "Instructions"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "doctor_id",
          "medication",
          "dosage",
          "instructions",
          "issued_at",
          "status"
        ],
        "title": "Prescription"
      },
      "PrescriptionCreateRequest": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "medication": {
            "type": "string",
            "minLength": 1,
            "title": "Medication"
          },
          "dosage": {
            "type": "string",
            "minLength": 1,
            "title": "Dosage"
          },
          "instructions": {
            "type": "string",
            "minLength": 1,
            "title": "Instructions"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "medication",
          "dosage",
          "instructions"
        ],
        "title": "PrescriptionCreateRequest"
      },
      "PrescriptionCreateResponse": {
        "properties": {
          "prescription": {
            "$ref": "#/components/schemas/Prescription"
          }
        },
        "type": "object",
        "required": [
          "prescription"
        ],
        "title": "PrescriptionCreateResponse"
      },
      "PrescriptionListResponse": {
        "properties": {
          "prescriptions": {
            "items": {
              "$ref": "#/components/schemas/Prescription"
            },
            "type": "array",
            "title": "Prescriptions"
          }
        },
        "type": "object",
        "required": [
          "prescriptions"
        ],
        "title": "PrescriptionListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Supabase session JWT",
        "scheme": "bearer"
      }
    }
  },
...
}